Move web service endpoint for getting an existing form's submission PDF from formio into help route
