+++ app/server/app/routes/index.js
@@ -5,7 +5,6 @@ const router = express.Router();
 router.use("/", require("./auth"));
 router.use("/api/content", require("./content"));
 router.use("/api/formio/nces", require("./formioNCES"));
-router.use("/api/formio/pdf", require("./formioPDF"));
 router.use("/api/user", require("./user"));
 router.use("/api/config", require("./config"));
 router.use("/api/bap", require("./bap"));
